+++ docs/django_settings.py
@@ -1,0 +1,9 @@
+# Dummy Django settings
+SECRET_KEY = 'dummy-key'
+
+from rstview.settings import *
+
+INSTALLED_APPS = (
+    'django.contrib.staticfiles',
+    'rstview',
+)